+++ src/order/application/orders.service.ts
@@ -2,11 +2,11 @@ import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
 import { Order } from "../domain/order.entity";
-import { OrderNotFoundException } from "../exception/OrderNotFoundException.exception";
-import { IOrderService } from "./order.service.interface";
+import { OrderNotFoundException } from "../exception/OrdersNotFoundException.exception";
+import { IOrderService } from "./orders.service.interface";
 
 @Injectable()
-export class OrderService implements IOrderService {
+export class OrdersService implements IOrderService {
   constructor(
     @InjectRepository(Order)
     private readonly orderRepository: Repository<Order>,
